// web/src/__generated/model/dto/AccountDto.ts
import type {AccountPreferences} from '../static/';

export type AccountDto = {
    'AccountController/DEFAULT_ACCOUNT_FETCHER': {
        readonly id: number;
        readonly name: string;
        readonly email: string;
        readonly admin: boolean;
        readonly preferences: AccountPreferences;
    }
}